requirements.txt
aiogram>=3.0.0
icalendar==5.0.7
recurring-ical-events==2.0.2
aiohttp>=3.8.5
python-dateutil==2.8.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1